[FEAT] Store Boosts Used in GameState
- Create universal function to update boostUsedAt times
- Harvest Greenhouse
- Harvest Crop Machine
- Harvest Fruit
- Harvest Crops
- Various Calendar Events
