+++ src/features/game/events/landExpansion/harvestGreenHouse.ts
@@ -1,4 +1,8 @@
-import { CriticalHitName, GameState } from "features/game/types/game";
+import {
+  BoostName,
+  CriticalHitName,
+  GameState,
+} from "features/game/types/game";
 import { isGreenhouseCrop, MAX_POTS } from "./plantGreenhouse";
 import {
   GREENHOUSE_CROPS,
@@ -15,7 +19,8 @@ import {
 } from "features/game/types/bumpkinActivity";
 import { produce } from "immer";
 import { getFruitYield } from "./fruitHarvested";
-import { getCropYieldAmount } from "./plant";
+import { getCropYieldAmount } from "./harvest";
+import { updateBoostUsed } from "features/game/types/updateBoostUsed";
 
 export const GREENHOUSE_CROP_TIME_SECONDS: Record<
   GreenHouseCropName | GreenHouseFruitName,
@@ -48,15 +53,15 @@ export function getGreenhouseCropYieldAmount({
   game: GameState;
   createdAt: number;
   criticalDrop?: (name: CriticalHitName) => boolean;
-}): number {
+}): { amount: number; boostsUsed: BoostName[] } {
   if (isGreenhouseCrop(crop)) {
-    const { amount } = getCropYieldAmount({
+    const { amount, boostsUsed } = getCropYieldAmount({
       crop,
       game,
       criticalDrop,
       createdAt,
     });
-    return amount;
+    return { amount, boostsUsed };
   }
 
   return getFruitYield({ name: crop, game, criticalDrop });
@@ -107,14 +112,14 @@ export function harvestGreenHouse({
     }
 
     // Harvests Crop
-    const greenhouseProduce =
-      pot.plant.amount ??
-      getGreenhouseCropYieldAmount({
-        crop: pot.plant.name,
-        game,
-        createdAt,
-        criticalDrop: (name) => !!(pot.plant?.criticalHit?.[name] ?? 0),
-      });
+    const { amount: greenhouseProduce, boostsUsed } = pot.plant.amount
+      ? { amount: pot.plant.amount, boostsUsed: [] }
+      : getGreenhouseCropYieldAmount({
+          crop: pot.plant.name,
+          game,
+          createdAt,
+          criticalDrop: (name) => !!(pot.plant?.criticalHit?.[name] ?? 0),
+        });
 
     const previousAmount = game.inventory[pot.plant.name] ?? new Decimal(0);
     game.inventory[pot.plant.name] = previousAmount.add(greenhouseProduce);
@@ -123,6 +128,12 @@ export function harvestGreenHouse({
     const activityName: BumpkinActivityName = `${pot.plant.name} Harvested`;
 
     game.bumpkin.activity = trackActivity(activityName, game.bumpkin.activity);
+
+    game.boostsUsedAt = updateBoostUsed({
+      game,
+      boostNames: boostsUsed,
+      createdAt,
+    });
 
     // Clears Pot
     delete pot.plant;
