+++ src/features/island/buildings/components/building/cropMachine/CropMachineModal.tsx
@@ -259,7 +259,9 @@ export const CropMachineModalContent: React.FC<Props> = ({
   ];
 
   const allowedSeeds = ALLOWED_SEEDS(state.bumpkin, inventory);
-  const cropYield = selectedPack ? getPackYieldAmount(state, selectedPack) : 0;
+  const cropYield = selectedPack
+    ? getPackYieldAmount(state, selectedPack).amount
+    : 0;
 
   return (
     <CloseButtonPanel
